requirements.txt
awscurl==0.36
blinker==1.9.0
boto3==1.42.14
botocore~=1.42.14
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
ConfigArgParse==1.7.1
configparser==7.2.0
Flask==3.1.2
Flask-Limiter==3.5.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
s3transfer==0.16.0
six==1.17.0
urllib3==2.6.2
Werkzeug==3.1.4
WTForms==3.1.2